refactor primitives for abi embedding
