+++ Cargo.toml
@@ -16,4 +16,4 @@ schemars = "0.8.8"
 serde_json = "1"
 
 [features]
-versioned-entries = []
+chunked-entries = []
